Locking Down DeleteItem Permissions

+++ backend/src/resolvers/Mutation.js
@@ -39,7 +39,19 @@ const Mutations = {
   },
   async deleteItem(_parent, args, ctx, info) {
     const where = { id: args.id };
-    const item = await ctx.db.query.item({ where }, `{ id, title }`);
+    const item = await ctx.db.query.item(
+      { where },
+      `{ id, title, user { id } }`
+    );
+
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some((permission) =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error('You do not have permission to do that!');
+    }
 
     return ctx.db.mutation.deleteItem({ where }, info);
   },
